+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/AbstractToggleMenuItemBuilder.java
@@ -22,13 +22,13 @@ import is.codion.common.state.ObservableState;
 import is.codion.common.state.State;
 import is.codion.swing.common.ui.component.button.DefaultToggleButtonBuilder.SetComponentValue;
 import is.codion.swing.common.ui.component.value.ComponentValue;
-import is.codion.swing.common.ui.control.Control;
 import is.codion.swing.common.ui.control.ToggleControl;
 
 import javax.swing.JMenuItem;
 import javax.swing.SwingConstants;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Supplier;
 
 import static java.util.Objects.requireNonNull;
 
@@ -53,8 +53,8 @@ abstract class AbstractToggleMenuItemBuilder<C extends JMenuItem, B extends Togg
 	}
 
 	@Override
-	public final B toggle(Control.Builder<ToggleControl, ?> toggleControl) {
-		return toggle(requireNonNull(toggleControl).build());
+	public final B toggle(Supplier<ToggleControl> toggleControl) {
+		return toggle(requireNonNull(toggleControl).get());
 	}
 
 	@Override